refactor(gsd): extract auto-worktree session registry
